cases in testfile, renamed file
